# source/asr-candidate/docs/api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "云蝠 Voice Runtime 管理与观测接口",
    "version": "1.13.0",
    "description": "云蝠 Voice Runtime 管理API：保留FreeSWITCH迁移对照，优先面向语音智能体媒体链路。1.10新增真实G711处理能力状态、显式隔离压测模式和完整内容/采样验收。未实现能力保持明确边界，不宣称完整FreeSWITCH或生产万路认证。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:9080",
      "description": "本机示例；实际由active.admin.listen决定。"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/v1/config": {
      "get": {
        "operationId": "getConfig",
        "summary": "读取运行配置和待启动草稿",
        "description": "读取共享revision、当前进程CSRF令牌、active/desired、持久化状态和Guard。可选ASR配置为部署原始值，未启用时省略；有效默认值见ASRStreamConfig。",
        "tags": [
          "运行管理"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminConfigResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "x-freeswitch": {
          "interface": "XML配置 / reloadxml",
          "relationship": "analogous_not_wire_compatible",
          "difference": "配置为RustSwitch JSON，不执行FreeSWITCH XML或reloadxml。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a/conf/vanilla"
        }
      },
      "put": {
        "operationId": "putConfig",
        "summary": "保存完整待启动配置",
        "description": "先校验完整config再检查revision；程序、日志、放音根目录、XML拨号计划、PCM流入口和ASR代理由部署启动文件管理，须保持原对象及启用状态，改变会返回422。当前Guard必须适合新的硬上限。其他设置保存不切换当前socket/worker，下次启动应用desired。",
        "tags": [
          "运行管理"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminConfigResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMedia"
          },
          "400": {
            "$ref": "#/components/responses/BadJSON"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation",
            "description": "配置或冻结字段校验失败；包括ASR目标格式、预算、路径冲突，以及改变ASR部署对象/启用状态。具体错误使用既有Validation响应。"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "XML配置 / reloadxml",
          "relationship": "analogous_not_wire_compatible",
          "difference": "保存下次启动参数，不是reloadxml的运行语义。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a/conf/vanilla"
        },
        "security": [
          {
            "LocalCSRF": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "单个JSON对象；Go重复已知键采用后值覆盖/对象合并，省略按零值，不是PATCH。",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdate"
              }
            }
          }
        }
      }
    },
    "/v1/guard": {
      "get": {
        "operationId": "getGuard",
        "summary": "读取即时峰值保护",
        "description": "GET会结算令牌补充，不消费额度或增加准入计数。",
        "tags": [
          "运行管理"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "x-freeswitch": {
          "interface": "fsctl max_sessions / fsctl sps",
          "relationship": "analogous_not_wire_compatible",
          "difference": "这是RustSwitch自有HTTP契约，不兼容原版ESL/fs_cli的传输、命令与输出。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        }
      },
      "put": {
        "operationId": "putGuard",
        "summary": "原子更新即时峰值保护",
        "description": "按revision检查，并同时受active/desired硬边界约束；先持久化再应用。降限不挂断已有通话，更新和启停不补满令牌。",
        "tags": [
          "运行管理"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMedia"
          },
          "400": {
            "$ref": "#/components/responses/BadJSON"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "x-freeswitch": {
          "interface": "fsctl max_sessions / fsctl sps",
          "relationship": "analogous_not_wire_compatible",
          "difference": "采用桥接资源数，新增建立中限制和软降速，不等同FreeSWITCH session计数或fsctl语法。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        },
        "security": [
          {
            "LocalCSRF": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "单个JSON对象；Go重复已知键采用后值覆盖/对象合并，省略按零值，不是PATCH。",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardUpdate"
              }
            }
          }
        }
      }
    },
    "/v1/drain": {
      "post": {
        "operationId": "drain",
        "summary": "暂停新初始呼叫",
        "description": "只修改当前draining，不持久化、不推进revision、不挂断存量。handler不解析正文；空/未知/非JSON正文不会因正文被拒绝，但仍要求Content-Type和CSRF。",
        "tags": [
          "运行管理"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/Origin"
          },
          {
            "$ref": "#/components/parameters/JSONContentType"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrainState"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMedia"
          }
        },
        "x-freeswitch": {
          "interface": "fsctl pause",
          "relationship": "analogous_not_wire_compatible",
          "difference": "只控制RustSwitch新初始INVITE，不区分FreeSWITCH inbound/outbound，也不是优雅停机命令。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        },
        "security": [
          {
            "LocalCSRF": []
          }
        ],
        "requestBody": {
          "required": false,
          "description": "正文不读取；推荐{}，不要求revision。",
          "content": {
            "application/json": {
              "schema": {},
              "example": {}
            }
          }
        }
      }
    },
    "/v1/resume": {
      "post": {
        "operationId": "resume",
        "summary": "恢复按既有策略接入",
        "description": "只修改当前draining，不持久化、不推进revision、不挂断存量。handler不解析正文；空/未知/非JSON正文不会因正文被拒绝，但仍要求Content-Type和CSRF。stopping=true时409。",
        "tags": [
          "运行管理"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/Origin"
          },
          {
            "$ref": "#/components/parameters/JSONContentType"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrainState"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMedia"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "x-freeswitch": {
          "interface": "fsctl resume",
          "relationship": "analogous_not_wire_compatible",
          "difference": "只控制RustSwitch新初始INVITE，不区分FreeSWITCH inbound/outbound，也不是优雅停机命令。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/blob/ef32e205295e29f034f1453ad245ba5efb07b94a/src/mod/applications/mod_commands/mod_commands.c"
        },
        "security": [
          {
            "LocalCSRF": []
          }
        ],
        "requestBody": {
          "required": false,
          "description": "正文不读取；推荐{}，不要求revision。",
          "content": {
            "application/json": {
              "schema": {},
              "example": {}
            }
          }
        }
      }
    },
    "/v1/fs-config": {
      "get": {
        "operationId": "getFSConfig",
        "summary": "读取FreeSWITCH配置草稿目录",
        "description": "基线叠加保存改动后的文件和参数，仅编辑导出，不是模块运行清单。",
        "tags": [
          "配置草稿"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FSConfigResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "conf/vanilla XML",
          "relationship": "export_artifact_only",
          "difference": "配置编辑与导出可用，RustSwitch不执行XML业务语义。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a/conf/vanilla"
        }
      },
      "put": {
        "operationId": "putFSConfig",
        "summary": "按参数ID保存差量修改",
        "description": "只发送本次ID到值的差量，XML字符自动转义。未知ID返回422；省略/null/空映射不改XML但仍推进revision。",
        "tags": [
          "配置草稿"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FSConfigResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMedia"
          },
          "400": {
            "$ref": "#/components/responses/BadJSON"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "conf/vanilla XML属性",
          "relationship": "export_artifact_only",
          "difference": "这是RustSwitch自有HTTP契约，不兼容原版ESL/fs_cli的传输、命令与输出。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a/conf/vanilla"
        },
        "security": [
          {
            "LocalCSRF": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "单个JSON对象；Go重复已知键采用后值覆盖/对象合并，省略按零值，不是PATCH。",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FSParametersUpdate"
              }
            }
          }
        }
      }
    },
...
    "/v1/fs-config/export": {
      "get": {
        "operationId": "getFSExport",
        "summary": "导出完整XML配置ZIP",
        "description": "包含conf/树、LICENSE.freeswitch、SOURCE.json和README-中文.txt。下载不部署或重载。合并集合失败返回500；ZIP构建错误当前handler直接返回，可能得到空/不完整响应，不能保证结构化500。",
        "tags": [
          "配置草稿"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string",
                  "description": ""
                },
                "example": "attachment; filename=\"freeswitch-1.11.3-config.zip\""
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "description": "非负累计数或即时数量。",
                  "minimum": 0
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "conf/vanilla配置分发",
          "relationship": "export_artifact_only",
          "difference": "这是RustSwitch自有HTTP契约，不兼容原版ESL/fs_cli的传输、命令与输出。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a/conf/vanilla"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminConfigResponse": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "integer",
            "description": "全部配置页共享的乐观并发版本，必须使用最近 GET 返回值。",
            "minimum": 1,
            "maximum": 18446744073709551615
          },
          "csrf_token": {
            "type": "string",
            "description": "本进程写令牌，重启后更换；不是远程身份认证。"
          },
          "active": {
            "$ref": "#/components/schemas/Config"
          },
          "desired": {
            "$ref": "#/components/schemas/Config"
          },
          "restart_required": {
            "type": "boolean",
            "description": "仅比较 active/desired，不包含 Guard 或 XML。"
          },
          "guard": {
            "$ref": "#/components/schemas/GuardSnapshot"
          },
          "persisted": {
            "type": "boolean",
            "description": "已经载入或保存管理状态，不表示草稿已运行。"
          },
          "notice": {
            "type": "string",
            "description": "启动优先级等提示，允许空串。"
          }
        },
        "description": "",
        "required": [
          "revision",
          "csrf_token",
          "active",
          "desired",
          "restart_required",
          "guard",
          "persisted",
          "notice"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Forbidden": {
        "description": "Host不是localhost/回环字面量，或写请求Origin/CSRF校验失败。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "parameters": {
      "Host": {
        "name": "Host",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "description": ""
        },
        "description": "客户端通常自动设置。主机部分须为精确localhost或回环IP字面量，自定义域名即便解析到回环也拒绝。"
      },
...
    },
    "securitySchemes": {
      "LocalCSRF": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RustSwitch-CSRF",
        "description": "读取/v1/config.csrf_token；重启更换，仅防跨站写入，不是远程认证。"
      }
    }
  },
...
}